modify: change id number datatype from int to bigint

+++ migrations/20240513144215_create_medical_patients_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS medical_patients (
     id VARCHAR(26) PRIMARY KEY,
     created_at TIMESTAMPTZ DEFAULT now(),
-    identity_number INT NOT NULL,
+    identity_number BIGINT NOT NULL,
     phone_number VARCHAR(15) NOT NULL,
     name VARCHAR(30) NOT NULL,
     birth_date VARCHAR NOT NULL,
